[#MOB-1105] Voting rewrite phase 3a: extract RoundListItem to top level
- New `Models/RoundListItem.swift` — extracted from `Voting.State.RoundListItem` so both the legacy reducer and the new `VotingCoordFlow` can reference the same type during the migration.
- Removed the nested definition from `Voting.State`.
- Updated all references in VotingStore+Session, PollsListView, and the voting unit tests from `Voting.State.RoundListItem` to top-level `RoundListItem`.
- No behavior change. Legacy voting flow continues to work.
